fix: address Sonar and Codacy findings
- Rename the Seek parameter to `origin`, matching the `Stream` declaration
- Remove two redundant int casts in `CopyOnWriteViewBacking`
- Convert the test structs to record structs, which implement `IEquatable<T>`
- Suppress S2325 on the `MemoryMappedFile` extension property: a C# extension
  property cannot be static (same false positive as the suppressed CA1822)
- Suppress S3874 on `Read<T>`/`Write<T>` of `IMemoryMappedViewAccessor`: the
  out/ref modifiers deliberately mirror the BCL `UnmanagedMemoryAccessor`
  signatures, so user code works unchanged against the real and mocked API

+++ Source/Testably.Abstractions.MemoryMappedFiles/CopyOnWriteViewBacking.cs
@@ -41,7 +41,7 @@ internal sealed class CopyOnWriteViewBacking(MemoryMappedViewBacking shared, lon
 			{
 				long pageIndex = (position + read) / PageSize;
 				int offsetInPage = (int)((position + read) % PageSize);
-				int chunk = (int)Math.Min(count - read, PageSize - offsetInPage);
+				int chunk = Math.Min(count - read, PageSize - offsetInPage);
 				if (_privatePages.TryGetValue(pageIndex, out byte[]? page))
 				{
 					Array.Copy(page, offsetInPage, buffer, offset + read, chunk);
@@ -68,7 +68,7 @@ internal sealed class CopyOnWriteViewBacking(MemoryMappedViewBacking shared, lon
 			{
 				long pageIndex = (position + written) / PageSize;
 				int offsetInPage = (int)((position + written) % PageSize);
-				int chunk = (int)Math.Min(count - written, PageSize - offsetInPage);
+				int chunk = Math.Min(count - written, PageSize - offsetInPage);
 				byte[] page = GetOrCreatePrivatePage(pageIndex);
 				Array.Copy(buffer, offset + written, page, offsetInPage, chunk);
 				written += chunk;
